requirements.txt
kaggle
numpy
pandas
statsmodels
scikit-learn
seaborn
catboost
streamlit
psycopg2-binary
sqlalchemy
scikit-surprise
